+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/application/dto/command/SearchGatheringCommand.java
@@ -1,7 +1,11 @@
 package com.sparta.moim.gathering.gathering.application.dto.command;
 
+import com.sparta.moim.common.security.CustomUserDetails;
 import com.sparta.moim.gathering.gathering.domain.dto.criteria.SearchGatheringCriteria;
+import com.sparta.moim.gathering.gathering.domain.entity.Member;
+import com.sparta.moim.gathering.gathering.domain.enums.MemberType;
 import java.time.LocalDateTime;
+import java.util.List;
 import java.util.UUID;
 import lombok.Builder;
 
@@ -32,5 +36,34 @@ public record SearchGatheringCommand(
         .size(size)
         .sort(sort)
         .build();
+  }
+
+  public static record JoinGatheringCommand(UUID gatheringId, String username) {
+
+    public JoinGatheringCommand(UUID gatheringId, CustomUserDetails userInfo) {
+      this(gatheringId, userInfo.getUsername());
+    }
+
+    public Member toDomain() {
+      return Member.builder()
+          .gatheringId(gatheringId)
+          .memberId(username)
+          .type(MemberType.GENERAL)
+          .joinTime(LocalDateTime.now())
+          .build();
+    }
+  }
+
+  public static record LeaveGatheringCommand(UUID gatheringId, String username) {
+    public LeaveGatheringCommand(UUID gatheringId, CustomUserDetails userInfo) {
+      this(gatheringId, userInfo.getUsername());
+    }
+  }
+
+  public static record RemoveGatheringCommand(
+      UUID gatheringId,
+      List<String> users,
+      String memberId
+  ) {
   }
 }
