+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/presentation/contoller/external/MemberController.java
@@ -1,9 +1,9 @@
-package com.sparta.moim.gathering.member.presentation.controller.external;
+package com.sparta.moim.gathering.gathering.presentation.contoller.external;
 
-import com.sparta.moim.gathering.member.application.dto.command.JoinGatheringCommand;
-import com.sparta.moim.gathering.member.application.dto.command.LeaveGatheringCommand;
-import com.sparta.moim.gathering.member.application.service.MemberService;
-import com.sparta.moim.gathering.member.presentation.dto.request.RemoveGatheringRequest;
+import com.sparta.moim.gathering.gathering.application.dto.command.SearchGatheringCommand.JoinGatheringCommand;
+import com.sparta.moim.gathering.gathering.application.dto.command.SearchGatheringCommand.LeaveGatheringCommand;
+import com.sparta.moim.gathering.gathering.application.service.MemberService;
+import com.sparta.moim.gathering.gathering.presentation.dto.request.RemoveGatheringRequest;
 import com.sparta.moim.common.response.ApiResponseData;
 import com.sparta.moim.common.security.CustomUserDetails;
 import jakarta.validation.Valid;
