+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/presentation/dto/response/GetMemberListResponse.java
@@ -1,4 +1,4 @@
-package com.sparta.moim.gathering.member.presentation.dto.response;
+package com.sparta.moim.gathering.gathering.presentation.dto.response;
 
 public record GetMemberListResponse(
     String name,
