--- gathering-service/src/main/java/com/sparta/moim/gathering/member/application/dto/command/LeaveGatheringCommand.java
@@ -1,10 +1,0 @@
-package com.sparta.moim.gathering.member.application.dto.command;
-
-import com.sparta.moim.common.security.CustomUserDetails;
-import java.util.UUID;
-
-public record LeaveGatheringCommand(UUID gatheringId, String username) {
-  public LeaveGatheringCommand(UUID gatheringId, CustomUserDetails userInfo) {
-    this(gatheringId, userInfo.getUsername());
-  }
-}
