--- gathering-service/src/main/java/com/sparta/moim/gathering/member/application/dto/command/RemoveGatheringCommand.java
@@ -1,11 +1,0 @@
-package com.sparta.moim.gathering.member.application.dto.command;
-
-import java.util.List;
-import java.util.UUID;
-
-public record RemoveGatheringCommand(
-    UUID gatheringId,
-    List<String> users,
-    String memberId
-) {
-}
