--- gathering-service/src/main/java/com/sparta/moim/gathering/member/domain/enums/MemberType.java
@@ -1,5 +1,0 @@
-package com.sparta.moim.gathering.member.domain.enums;
-
-public enum MemberType {
-  ADMIN, GENERAL
-}
